security(uds): keep the socket-path fallback in a verified per-user directory
When {run_dir}/instances/{id}.sock is longer than the platform's sun_path
limit, the socket moves to a short hashed path in /tmp. That directory was
/tmp/capsem, created 0755 by whichever user came first and shared by everyone
after: under a world-writable /tmp any other user of the host could unlink a
service's socket or bind their own at the path clients were handed in the
provision response, and the VZ save/restore lock beside it could be a planted
symlink. The fallback is now /tmp/capsem-<uid>, created 0700 and refused
unless it is a real directory owned by the calling user with no group or
other bits; instance_socket_path and terminal_socket_path return that refusal
instead of a path, and every caller -- service, gateway, capsem-process and
the CLI -- propagates it. The VZ lock file lives in the same directory.

Pure moves so the two service files that grew stay under their ratchets:
ServiceState's session-dir housekeeping into session_housekeeping.rs and the
log/panic/triage routes into vm_files/diagnostics.rs.
