+++ config/gate.toml
@@ -1088,9 +1088,9 @@ must_stay_below_lines = 3000
 "crates/capsem-process/src/vsock.rs" = 1690
 "crates/capsem-proto/src/tests.rs" = 1591
 "crates/capsem-service/src/ledger_routes.rs" = 2931
-"crates/capsem-service/src/main.rs" = 2516
+"crates/capsem-service/src/main.rs" = 2309
 "crates/capsem-service/src/profile_routes.rs" = 2399
-"crates/capsem-service/src/vm_files.rs" = 2397
+"crates/capsem-service/src/vm_files.rs" = 2036
 "crates/capsem-service/src/vm_lifecycle.rs" = 1045
 "crates/capsem-service/src/tests/assets_registry.rs" = 1600
 "crates/capsem-service/src/tests/lifecycle.rs" = 2436
