+++ crates/capsem-foundation/src/uds.rs
@@ -2,40 +2,93 @@
 //!
 //! macOS caps `sockaddr_un.sun_path` at 104 bytes; Linux at 108. Temp dirs on
 //! macOS (`/var/folders/lv/…`) easily blow past this, so per-VM socket paths
-//! must fall back to a short hashed path under `/tmp/capsem/`.
+//! must fall back to a short hashed path under a private per-user directory
+//! in `/tmp`.
 //!
 //! This module is the single source of truth for that rule. Clients MUST NOT
 //! recompute the fallback path -- the fallback hash uses `DefaultHasher` which
 //! is not stable across processes. Callers get the chosen path from the
 //! service via the provision response.
 
+use std::io;
 use std::path::{Path, PathBuf};
 
 /// Maximum length of a UDS path we'll accept before falling back to
-/// `/tmp/capsem/<hash>.sock`. Chosen well under macOS's 104 and Linux's 108
+/// `/tmp/capsem-<uid>/<hash>.sock`. Chosen well under macOS's 104 and Linux's 108
 /// so there's headroom for any suffix.
 pub const SUN_PATH_MAX: usize = 90;
 
 /// Compute the UDS socket path for a VM instance.
 ///
 /// Returns `{run_dir}/instances/{id}.sock` when that fits under
-/// `SUN_PATH_MAX`; otherwise a short hashed path under `/tmp/capsem/`.
+/// `SUN_PATH_MAX`; otherwise a short hashed path under the private per-user
+/// fallback directory, which is an error when that directory cannot be
+/// created or is not ours alone.
 ///
 /// The hashed path uses `DefaultHasher` which is randomised per-process --
 /// so this function's output is ONLY valid in the process that originally
 /// computed it. Other processes must receive the chosen path via IPC.
-pub fn instance_socket_path(run_dir: &Path, id: &str) -> PathBuf {
+pub fn instance_socket_path(run_dir: &Path, id: &str) -> io::Result<PathBuf> {
     let preferred = run_dir.join("instances").join(format!("{id}.sock"));
     if preferred.as_os_str().len() < SUN_PATH_MAX {
-        return preferred;
+        return Ok(preferred);
     }
     use std::hash::{Hash, Hasher};
     let mut h = std::collections::hash_map::DefaultHasher::new();
     id.hash(&mut h);
     run_dir.hash(&mut h);
-    let dir = PathBuf::from("/tmp/capsem");
-    let _ = std::fs::create_dir_all(&dir);
-    dir.join(format!("{:x}.sock", h.finish()))
+    Ok(private_fallback_dir()?.join(format!("{:x}.sock", h.finish())))
+}
+
+/// The calling user's numeric id.
+pub fn current_uid() -> u32 {
+    // SAFETY: getuid(2) takes no arguments and cannot fail.
+    unsafe { libc::getuid() }
+}
+
+/// `/tmp/capsem-<uid>`, created 0700 and verified to be exactly that.
+///
+/// The fallback used to live in a shared `/tmp/capsem`, created 0755 by
+/// whichever user came first. Under a world-writable `/tmp` that directory
+/// belonged to somebody else for everyone after them: they could unlink a
+/// service's socket or bind their own at a path the service was about to
+/// use, and every client holding that path from the provision response would
+/// dial it. A directory that is a symlink, not ours, or readable by anyone
+/// else is refused rather than used.
+pub fn private_fallback_dir() -> io::Result<PathBuf> {
+    private_fallback_dir_under(Path::new("/tmp"))
+}
+
+fn private_fallback_dir_under(base: &Path) -> io::Result<PathBuf> {
+    use std::os::unix::fs::{DirBuilderExt, MetadataExt};
+    let uid = current_uid();
+    let dir = base.join(format!("capsem-{uid}"));
+    match std::fs::DirBuilder::new().mode(0o700).create(&dir) {
+        Ok(()) => {}
+        Err(error) if error.kind() == io::ErrorKind::AlreadyExists => {}
+        Err(error) => {
+            return Err(io::Error::new(
+                error.kind(),
+                format!("create private socket dir {}: {error}", dir.display()),
+            ))
+        }
+    }
+    let refuse = |what: String| io::Error::other(format!("refusing socket fallback dir {}: {what}", dir.display()));
+    let metadata = std::fs::symlink_metadata(&dir)?;
+    if metadata.file_type().is_symlink() {
+        return Err(refuse("it is a symlink".to_string()));
+    }
+    if !metadata.is_dir() {
+        return Err(refuse("it is not a directory".to_string()));
+    }
+    if metadata.uid() != uid {
+        return Err(refuse(format!("it is owned by uid {}, not {uid}", metadata.uid())));
+    }
+    let mode = metadata.mode() & 0o777;
+    if mode & 0o077 != 0 {
+        return Err(refuse(format!("its mode is {mode:o}, not 700")));
+    }
+    Ok(dir)
 }
 
 /// Compute the terminal WebSocket UDS path for a VM instance.
@@ -52,16 +105,16 @@ pub fn instance_socket_path(run_dir: &Path, id: &str) -> PathBuf {
 /// directory. Past that every connection failed with `path must be shorter
 /// than SUN_LEN`, logged at ERROR on each retry and surfaced to the user as a
 /// session whose shell never appeared.
-pub fn terminal_socket_path(run_dir: &Path, id: &str) -> PathBuf {
+pub fn terminal_socket_path(run_dir: &Path, id: &str) -> io::Result<PathBuf> {
     let preferred = run_dir.join("instances").join(format!("{id}-ws.sock"));
     if preferred.as_os_str().len() < SUN_PATH_MAX {
-        return ensured(preferred);
+        return Ok(ensured(preferred));
     }
     let mut digest = blake3::Hasher::new();
     digest.update(run_dir.as_os_str().as_encoded_bytes());
     digest.update(id.as_bytes());
     let short = &digest.finalize().to_hex()[..16];
-    ensured(PathBuf::from("/tmp/capsem").join(format!("{short}-ws.sock")))
+    Ok(private_fallback_dir()?.join(format!("{short}-ws.sock")))
 }
 
 /// A path with a directory to bind in.
